requirements.txt
librosa==0.9.2
numpy==1.23.3
pandas==1.4.2
matplotlib==3.3.4
seaborn
japanize-matplotlib==1.1.3
torch==2.0.0
torchvision
scienceplots==2.1.0
retry==0.9.2
pyyaml
miditok
miditoolkit==0.1.16
pretty_midi
python-rtmidi==1.4.9
scipy==1.9.3
scikit-learn
streamlit-option-menu